+++ mod-record-specifications-dto/src/main/java/org/folio/rspec/domain/dto/ValidationError.java
@@ -1,0 +1,17 @@
+package org.folio.rspec.domain.dto;
+
+import java.util.UUID;
+import lombok.Builder;
+import lombok.Value;
+
+@Value
+@Builder
+public class ValidationError {
+
+  String path;
+  SeverityType severity;
+  DefinitionType definitionType;
+  UUID definitionId;
+  String message;
+  String ruleCode;
+}