+++ mod-record-specifications-validator/src/main/java/org/folio/rspec/validation/converter/Converter.java
@@ -1,0 +1,6 @@
+package org.folio.rspec.validation.converter;
+
+public interface Converter<S, T> {
+
+  T convert(S source);
+}