+++ mod-record-specifications-validator/src/main/java/org/folio/rspec/validation/validator/SpecificationRuleCode.java
@@ -1,0 +1,6 @@
+package org.folio.rspec.validation.validator;
+
+public interface SpecificationRuleCode {
+
+  String getCode();
+}