+++ mod-record-specifications-validator/src/main/java/org/folio/rspec/validation/validator/marc/model/MarcField.java
@@ -1,0 +1,12 @@
+package org.folio.rspec.validation.validator.marc.model;
+
+public interface MarcField {
+
+  Reference reference();
+
+  boolean hasValue();
+
+  default String tag() {
+    return reference().getTag();
+  }
+}